# src/main/resources/static/openapi/inventory-api-v1.yaml
openapi: 3.0.3
info:
  title: PMIS  — Inventory API
  description: |
    REST API for Inventory Management (Safaricom Healthcare PMIS). Operational deployment
    aligns with the API gateway prefix **/api/v1**; `servers[0].url` should be appended
    before each path (for example `/inventory/products`).
  version: 1.1.0
  contact:
    name: PMIS Engineering
servers:
  - url: https://api.pmis.safaricom.com/api/v1
    description: Production (gateway)
  - url: https://api-staging.pmis.safaricom.com/api/v1
    description: Staging (gateway)
  - url: http://localhost:{port}/api/v1
    description: Local inventory service
    variables:
      port:
        default: "8080"
...
security:
  - BearerAuth: []
paths:
  /inventory/products:
    get:
      tags: [Products]
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/ProductCategory'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: manufacturer_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
    post:
      tags: [Products]
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '409':
          description: Duplicate PPB/NDC
  /inventory/products/drafts:
    post:
      tags: [Products]
      operationId: saveProductDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDraftRequest'
      responses:
        '201':
          description: Draft saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDraft'
  /inventory/products/drafts/{draft_id}:
    get:
      tags: [Products]
      operationId: getProductDraft
      parameters:
        - $ref: '#/components/parameters/DraftIdParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDraft'
    patch:
      tags: [Products]
      operationId: updateProductDraft
      parameters:
        - $ref: '#/components/parameters/DraftIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDraftRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDraft'
  /inventory/products/{product_id}:
    get:
      tags: [Products]
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
    patch:
      tags: [Products]
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
    delete:
      tags: [Products]
      operationId: deleteProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '204':
          description: Deleted
...
  /terminology/search:
    get:
      tags: [Terminology Service]
      operationId: searchTerminology
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: manufacturer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: source
          in: query
          schema:
            type: string
            enum: [rxnorm, ppb_ndc, all]
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminologySearchResponse'
  /terminology/products/{terminology_id}:
    get:
      tags: [Terminology Service]
      operationId: getTerminologyProduct
      parameters:
        - name: terminology_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminologyProduct'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductIdParam:
      name: product_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DraftIdParam:
      name: draft_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
    ProductCategory:
      type: string
      enum:
        - Antibiotics
        - Analgesics
        - Diabetes
        - Cardiovascular
        - Gastrointestinal
        - Antifungals
        - Antivirals
        - Vitamins
        - Other
    ProductStatus:
      type: string
      enum:
        [draft, available, low_stock, out_of_stock, controlled, recalled, expiring_soon, unknown]
...
    Pagination:
      type: object
      properties:
        page: {type: integer}
        limit: {type: integer}
        total: {type: integer}
        total_pages: {type: integer}
...
    DraftData:
      description: >-
        Partial **CreateProductRequest** snapshot; omit properties not yet populated.
        Mirrors CreateProduct field names (snake_case).
      type: object
      additionalProperties: true
    ProductDraftRequest:
      type: object
      required:
        - wizard_step
      properties:
        wizard_step:
          type: integer
          enum: [1, 2, 3, 4]
        data:
          $ref: '#/components/schemas/DraftData'
    CreateProductRequest:
      type: object
      properties:
        product_name: {type: string}
        generic_name: {type: string}
        manufacturer_id: {type: string, format: uuid, nullable: true}
        category:
          $ref: '#/components/schemas/ProductCategory'
    ProductDraft:
      type: object
      properties:
        id: {type: string, format: uuid}
        wizard_step: {type: integer, enum: [1,2,3,4]}
        data:
          $ref: '#/components/schemas/CreateProductRequest'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateProductRequest:
      type: object
      additionalProperties: true
    ProductSummary:
      type: object
      additionalProperties: true
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          additionalProperties: true
    ProductListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    TerminologySearchResponse:
      type: object
      additionalProperties: true
    TerminologyProduct:
      type: object
      additionalProperties: true
...